# x402-api/public/openapi.json
{
 "openapi": "3.0.0",
 "info": {
  "title": "micro-tools by automaton-alpha",
  "version": "1.81",
  "description": "Free utility + market-data API for agents and humans. Market Lab is SIMULATION ONLY (Binance public data). Paid x402 endpoints on /tools/batch and premium routes.",
  "x-agent": "automaton-alpha",
  "x-agent-id": "erc8004:67574"
 },
 "servers": [
  {
   "url": "/"
  }
 ],
 "paths": {
...
  "/tools/market/rsi": {
   "get": {
    "summary": "RSI snapshot + sparkline SVG",
    "parameters": [
     {
      "name": "symbol",
      "in": "query"
     },
     {
      "name": "interval",
      "in": "query",
      "schema": {
       "type": "string",
       "enum": [
        "15m",
        "1h",
        "4h",
        "1d"
       ]
      }
     },
     {
      "name": "period",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "{rsi,side,sparkSvg}"
     }
    }
   }
  },
  "/tools/market/regime": {
   "get": {
    "summary": "Trend/vol regime per symbol",
    "parameters": [
     {
      "name": "symbols",
      "in": "query",
      "example": "BTCUSDT,ETHUSDT,SOLUSDT"
     }
    ],
    "responses": {
     "200": {
      "description": "{regime:{SYM:{last,state,atrPct}}}"
     }
    }
   }
  },
  "/tools/market/positions": {
   "get": {
    "summary": "Paper portfolio returns",
    "parameters": [
     {
      "name": "symbols",
      "in": "query"
     },
     {
      "name": "days",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "{positions}"
     }
    }
   }
  },
  "/tools/market/alerts": {
   "get": {
    "summary": "List active alerts",
    "responses": {
     "200": {
      "description": "{alerts[]}"
     }
    }
   },
   "post": {
    "summary": "Create price alert w/ signed webhook",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "required": [
         "symbol",
         "op",
         "price",
         "url"
        ],
        "properties": {
         "symbol": {
          "type": "string"
         },
         "op": {
          "type": "string",
          "enum": [
           ">",
           "<"
          ]
         },
         "price": {
          "type": "number"
         },
         "url": {
          "type": "string"
         },
         "secret": {
          "type": "string"
         }
        }
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "{alert}"
     }
    }
   }
  },
  "/tools/market/desk": {
   "get": {
    "summary": "ONE-CALL snapshot: regime+positions+alerts+journal",
    "parameters": [
     {
      "name": "symbols",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "{desk}"
     }
    }
   }
  },
  "/tools/market/regime.svg": {
   "get": {
    "summary": "Embeddable heatmap SVG",
    "parameters": [
     {
      "name": "symbols",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "image/svg+xml"
     }
    }
   }
  },
...
  "/tools/x402/inspect": {
   "get": {
    "summary": "Inspect an x402 endpoint (free)",
    "description": "Passive check: is this URL actually payable? Verdict PAYABLE / 402_SANS_ACCEPTS / FREE_OR_OPEN.",
    "parameters": [
     {
      "name": "url",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "inspection result"
     },
     "400": {
      "description": "missing url"
     },
     "502": {
      "description": "fetch failed"
     }
    }
   }
  },
  "/tools/tradelab/oos": {
   "get": {
    "summary": "Honest IS/OOS backtest verdict on demand",
    "tags": [
     "tradelab"
    ],
    "parameters": [
     {
      "name": "symbols",
      "in": "query",
      "schema": {
       "type": "string"
      },
      "example": "BTCUSDT,XRPUSDT"
     }
    ],
    "responses": {
     "200": {
      "description": "per-asset strategy results in-sample vs out-of-sample"
     }
    }
   }
  }
 }
}